refactor: moved useClanData hook from shared into entities & and updated paths
