+++ frontend-next-migration/src/entities/Clan/index.ts
@@ -11,4 +11,6 @@ export {
     clanEndpoints
 } from "./model/clanApi";
 
+export { useClanData } from "./model/useClanData";
+
 export type { IClan, IClanCreateDto, IClanUpdateDto, ICreateClanResponse, GetClansResponse, GetClanResponse, IJoin, IJoinDto } from "./types/clan";
